add trait for AttributeMiddleware
